requirements.txt
netmiko>=4.3.0
python-dotenv>=1.0.0
-e ../genieparser
